Bug workaround for secured mode and listMethods.
A class in shindig DefaultServiceFetcher performs an internal http
request to listMethods. This request is not authenticated (no st=
parameter), so the request fails. This manifests as an error in the
console:

org.apache.shindig.gadgets.render.DefaultServiceFetcher retrieveServices
SEVERE: Failed to parse services methods from endpoint
http://localhost:8098/katari-sample/module/shindig/social/rpc.
JSONObject["data"] not found.

git-svn-id: http://katari.globant.com/svn/katari/main@43 e3f198be-754c-422e-81b5-761b7d6d5333